Move save file utils into repository
